Move filters and pages functions into downloader
