+++ minis/downloader/__init__.py
@@ -1,5 +1,5 @@
 from .downloader import *
-from .pages import *
-from .filters import *
+# from .pages import *
+# from .filters import *
 from .file_utils import *
 from .show_info import *
